move scraper logic to notebook
